test: skip kernel-native L7/XDP/DNS on unenhanced kernels
