+++ test/e2e/baseline_test.go
@@ -118,6 +118,7 @@ func OriginalSourceCheck(t framework.TestContext, src echo.Instance) echo.Checke
 // Test access to service, enabling L7 processing and propagating original src when  appropriate.
 func TestServices(t *testing.T) {
 	runTest(t, func(t framework.TestContext, src echo.Instance, dst echo.Instance, opt echo.CallOptions) {
+		requireEnhancedKernelForKernelNative(t)
 		if supportsL7(opt, src, dst) {
 			opt.Check = httpValidator
 		} else {
@@ -172,6 +173,7 @@ func TestPodIP(t *testing.T) {
 
 func TestDNSProxy(t *testing.T) {
 	runTest(t, func(t framework.TestContext, src echo.Instance, dst echo.Instance, opt echo.CallOptions) {
+		requireEnhancedKernelForKernelNative(t)
 		if opt.Scheme != scheme.HTTP {
 			return
 		}
@@ -239,6 +241,7 @@ spec:
 // This simulates external service access using DNS resolution.
 func TestServiceEntryDNSResolution(t *testing.T) {
 	runTest(t, func(t framework.TestContext, src echo.Instance, dst echo.Instance, opt echo.CallOptions) {
+		requireEnhancedKernelForKernelNative(t)
 		if opt.Scheme != scheme.HTTP {
 			return
 		}
